Split MadCoreCUDSS out of MadCoreCUDA
Move the cuDSS sparse solver (CUDSSSolver) and the GPU Schur-complement
KKT system (GPUSchurComplementKKTSystem) from MadCoreCUDA into a new
MadCoreCUDSS subpackage, so the cuSOLVER-based LapackCUDASolver / dense
GPU path no longer pulls in the heavy cuDSS library.

- New MadCore/lib/MadCoreCUDSS: cudss.jl, cuda_schur.jl, kernels_schur.jl
  moved from MadCoreCUDA; deps on MadCore + MadCoreCUDA + CUDA stack + CUDSS
  (+ AMD/Metis for the cuDSS reordering path).
- MadCoreCUDA: drop CUDSS dep, import, the three includes, and the CUDSS
  export/default_options; now CUDSS-free (LapackCUDASolver / dense GPU only).
- cuMadNLP: import CUDSSSolver from MadCoreCUDSS (was MadCoreCUDA); add
  MadCoreCUDSS dep/source/compat.
- CI test-gpu: Pkg.develop MadCoreCUDSS + coverage dir; CompatHelper subdir.

Validated on node 000 (Julia 1.11.9): MadCoreCUDSS precompiles and exports
CUDSSSolver + GPUSchurComplementKKTSystem; cuMadNLP precompiles and loads
(madnlp + CUDSSSolver present; GPU Schur create_kkt_system method registered
transitively via cuMadNLP -> MadCoreCUDSS).
