+++ MadCore/lib/MadCoreCUDSS/Project.toml
@@ -1,14 +1,16 @@
-name = "MadCoreCUDA"
-uuid = "bc4e0c13-4672-410c-a012-e283a446691e"
+name = "MadCoreCUDSS"
+uuid = "1f78e23c-6596-4cec-b75b-391af35fa428"
 version = "0.1.0"
 authors = ["MadNLP contributors"]
 
 [deps]
 AMD = "14f7f29c-3bd6-536c-9a0b-7339e30b5a3e"
 CUDACore = "bd0ed864-bdfe-4181-a5ed-ce625a5fdea2"
+CUDSS = "45b445bb-4962-46a0-9369-b4df9d0f772e"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MadCore = "6fc510d6-e066-472d-adc5-2ce5e9b14820"
+MadCoreCUDA = "bc4e0c13-4672-410c-a012-e283a446691e"
 MadCoreKernelAbstractions = "b5d449e2-bd97-4205-9c78-f5795562dab8"
 Metis = "2679e427-3c69-5b7f-982b-ece356f1e94b"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
@@ -18,17 +20,20 @@ cuSPARSE = "b26da814-b3bc-49ef-b0ee-c816305aa060"
 
 [sources]
 MadCore = {path = "../.."}
+MadCoreCUDA = {path = "../MadCoreCUDA"}
 MadCoreKernelAbstractions = {path = "../MadCoreKernelAbstractions"}
 
 [compat]
-SparseArrays = "1.10"
-LinearAlgebra = "1.10"
 AMD = "0.5"
 CUDACore = "6"
+CUDSS = "0.7, 0.8"
 KernelAbstractions = "0.9"
+LinearAlgebra = "1.10"
 MadCore = "0.1"
+MadCoreCUDA = "0.1"
 MadCoreKernelAbstractions = "0.1"
 Metis = "1"
+SparseArrays = "1.10"
 cuBLAS = "6"
 cuSOLVER = "6"
 cuSPARSE = "6"
