+++ cuMadNLP/Project.toml
@@ -8,12 +8,14 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 CUDACore = "bd0ed864-bdfe-4181-a5ed-ce625a5fdea2"
 GPUArraysCore = "46192b85-c4d5-4398-a991-12ede77f4527"
 MadCoreCUDA = "bc4e0c13-4672-410c-a012-e283a446691e"
+MadCoreCUDSS = "1f78e23c-6596-4cec-b75b-391af35fa428"
 MadCoreKernelAbstractions = "b5d449e2-bd97-4205-9c78-f5795562dab8"
 MadNLP = "2621e9c9-9eb4-46b1-8089-e8c72242dfb6"
 
 [sources]
 MadNLP = {path = ".."}
 MadCoreCUDA = {path = "../MadCore/lib/MadCoreCUDA"}
+MadCoreCUDSS = {path = "../MadCore/lib/MadCoreCUDSS"}
 MadCoreKernelAbstractions = {path = "../MadCore/lib/MadCoreKernelAbstractions"}
 
 [compat]
@@ -21,6 +23,7 @@ Reexport = "1"
 CUDACore = "6"
 GPUArraysCore = "0.2"
 MadCoreCUDA = "0.1"
+MadCoreCUDSS = "0.1"
 MadCoreKernelAbstractions = "0.1"
 MadNLP = "0.11"
 julia = "1.10"
